Inference/IoTSimulator/requirements.txt
pandas
azure.iot.device